--- a/usaspending_api/references/v2/views/filter_tree/filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/filter_tree.py
@@ -43,7 +43,7 @@ class FilterTree(metaclass=ABCMeta):
 
         retval = [
             self._linked_node_from_data(ancestor_array, elem, filter_string, child_layers)
-            for elem in self.raw_search(ancestor_array)
+            for elem in self.raw_search(ancestor_array, filter_string)
         ]
         if filter_string:
             retval = [elem for elem in retval if self.matches_filter(elem, filter_string)]
@@ -52,7 +52,7 @@ class FilterTree(metaclass=ABCMeta):
     def _linked_node_from_data(self, ancestor_array, data, filter_string, child_layers):
         retval = self.unlinked_node_from_data(ancestor_array, data)
         if child_layers:
-            raw_children = self.raw_search(ancestor_array + [retval.id])
+            raw_children = self.raw_search(ancestor_array + [retval.id], filter_string)
             temp_children = [
                 self._linked_node_from_data(ancestor_array + [retval.id], elem, filter_string, child_layers - 1)
                 for elem in raw_children