--- a/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
@@ -17,16 +17,16 @@ PSC_GROUPS = {
 
 
 class PSCFilterTree(FilterTree):
-    def raw_search(self, tiered_keys, filter_string=None):
+    def raw_search(self, tiered_keys):
         if not self._path_is_valid(tiered_keys):
             return []
 
         if len(tiered_keys) == 0:
             return self._toptier_search()
         elif len(tiered_keys) == 1:
-            return self._psc_from_group(tiered_keys[0], filter_string)
+            return self._psc_from_group(tiered_keys[0])
         else:
-            return self._psc_from_parent(tiered_keys[-1], filter_string)
+            return self._psc_from_parent(tiered_keys[-1])
 
     def _path_is_valid(self, path: list) -> bool:
         if len(path) > 1:
@@ -40,22 +40,18 @@ class PSCFilterTree(FilterTree):
     def _toptier_search(self):
         return PSC_GROUPS.keys()
 
-    def _psc_from_group(self, group, filter_string):
+    def _psc_from_group(self, group):
         # The default regex value will match nothing
         filters = [Q(code__iregex=PSC_GROUPS.get(group, {}).get("pattern") or "(?!)")]
-        # if filter_string:
-        #     filters.append(Q(Q(code__icontains=filter_string) | Q(description__icontains=filter_string)))
         return [{"id": object.code, "description": object.description} for object in PSC.objects.filter(*filters)]
 
-    def _psc_from_parent(self, parent, filter_string: str):
+    def _psc_from_parent(self, parent):
         # two out of three branches of the PSC tree "jump" over 3 character codes
         desired_len = len(parent) + 2 if len(parent) == 2 and parent[0] != "A" else len(parent) + 1
         filters = [
             Q(length=desired_len),
             Q(code__startswith=parent),
         ]
-        if filter_string and desired_len == 4:
-            filters.append(Q(Q(code__icontains=filter_string) | Q(description__icontains=filter_string)))
         return [{"id": object.code, "description": object.description} for object in PSC.objects.filter(*filters)]
 
     def unlinked_node_from_data(self, ancestors: list, data) -> UnlinkedNode:
@@ -63,3 +59,18 @@ class PSCFilterTree(FilterTree):
             return UnlinkedNode(id=data, ancestors=ancestors, description="")
         else:
             return UnlinkedNode(id=data["id"], ancestors=ancestors, description=data["description"])
+
+    def get_count(self, tiered_keys: list, id) -> int:
+        if len(tiered_keys) == 0:
+            filters = [Q(code__iregex=PSC_GROUPS.get(group, {}).get("pattern") or "(?!)")]
+            PSC.objects.filter(*filters).count()
+        if len(tiered_keys) == 1:
+            desired_len = len(parent) + 2 if len(parent) == 2 and parent[0] != "A" else len(parent) + 1
+            filters = [
+                Q(length=desired_len),
+                Q(code__startswith=parent),
+            ]
+            PSC.objects.filter(*filters).count()
+        if len(tiered_keys) == 2:
+            return 1
+        return 0