--- a/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
+++ b/usaspending_api/references/v2/views/filter_tree/psc_filter_tree.py
@@ -150,8 +150,8 @@ class PSCFilterTree(FilterTree):
                     "id": object.code,
                     "ancestors": ancestors,
                     "description": object.description,
-                    "count": self.get_count(ancestor_array, object.code),
-                    "children": [],
+                    "count": 0,
+                    "children": None,
                 }
             )
         return retval