--- a/usaspending_api/references/v2/views/filter_tree/tas.py
+++ b/usaspending_api/references/v2/views/filter_tree/tas.py
@@ -36,12 +36,5 @@ class TASViewSet(APIView):
 
         filter_tree = TASFilterTree()
         return Response(
-            {
-                "results": [
-                    elem.to_JSON()
-                    for elem in filter_tree.search(
-                        tier1, tier2, tier3, request_values["depth"], request_values["filter"]
-                    )
-                ]
-            }
+            {"results": filter_tree.search(tier1, tier2, tier3, request_values["depth"], request_values["filter"])}
         )