**Description:**
Completely rewrites the TAS/PSC filter endpoints that support the TAS/PSC filter selection on the advanced search page to improve performance.

**Technical details:**
The rewrite was necessary due to the old implementation - the new implementation is far less elegant, but the old one built out the tree completely just to get the counts for each node, which resulted in thousands for database queries just to return a few nodes, which wasn't great for performance.
The new implementation builds the tree from the bottom up, and does database queries for each level, then combine them in memory. 
Additionally fixes a bug in the Research and Development PSC nodes, where the `AU` subsection was not returning correctly due to the `AU` codes skipping the length 3 PSC codes, which other Research and development codes do not.
This PR also addresses bug [DEV-6092](https://federal-spending-transparency.atlassian.net/browse/DEV-6092), which was a result of the previous attempts to fix the performance. The 
**Requirements for PR merge:**

1. [X] Unit & integration tests updated
2. [X] API documentation updated
3. [ ] Necessary PR reviewers:
    - [ ] Backend
4. [N/A] Matview impact assessment completed
5. [X] Frontend impact assessment completed
6. [X] Data validation completed
7. [N/A] Appropriate Operations ticket(s) created
8. [X] Jira Ticket [DEV-5839](https://federal-spending-transparency.atlassian.net/browse/DEV-5839):
    - [X] Link to this Pull-Request
    - [X] Performance evaluation of affected (API | Script | Download)
    - [X] Before / After data comparison

**Area for explaining above N/A when needed:**
```
```
Performance test against sandbox:

  | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=073-1154 | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=salaries | {{api_url}}/api/v2/references/filter_tree/tas/ | {{api_url}}/api/v2/references/filter_tree/tas/091/ | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=stabilization | {{api_url}}/api/v2/references/filter_tree/tas/?depth=2&filter=institute |  
-- | -- | -- | -- | -- | -- | -- | --
sandbox (new) | 596 ms | 2.71 s | 1317 ms | 1121 ms | 716 ms | 2.81 s |  
sandbox (old) | 18.47 s | 19.29 s | 18.67 s | 1142 ms | 18.45 s | 19.48 s |  


  | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=research | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=education | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=salaries | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=EQPT | {{api_url}}/api/v2/references/filter_tree/psc/ | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=engineer | {{api_url}}/api/v2/references/filter_tree/psc/?depth=-1&filter=agriculture
-- | -- | -- | -- | -- | -- | -- | --
sandbox (new) | 895 ms | 475 ms | 577 ms | 636 ms | 578 ms | 855 ms | 698 ms
sandbox (old) | 1399 ms | 1012 ms | 978 ms | 886 ms | 4.30 s | 1424 ms | 982 ms