@sethstoudenmier 2020-09-18
Endpoint is still at a not ideal 20 seconds runtime against QAT, but it is a huge improvement! When testing locally the code changes in this PR brought the runtime down from 4 minutes to 20 seconds when first opening the TAS Filter. Subsequent choices to open Agencies and individual TAS are not an issue.